move download statistics to /middleware/sync_by_install.js. fix #747
